update_from() in text.py now also copies _antialiased
This ensures  the antialiasing of an axis is consistent.
The antialiasing of an axis is determined by
rcParams['text.antialiased'] when axis is created.
To be more specific, it's determined by the first Text
object created (internally by axis) when creating axis.

Currently there's no way to change after creation.